update to latest version of Greenwood and NodeJS
