+++ .nvmrc
@@ -1,1 +1,1 @@
-18.20.4
+22.14.0
